Refactorizes the handlers and implicits
