+++ core/src/main/scala/api/package.scala
@@ -17,8 +17,11 @@
 package freestyle.cassandra
 
 import cats.data.Kleisli
-import cats.~>
-import com.datastax.driver.core.{Cluster, ResultSet, Session}
+import cats.{~>, MonadError}
+import com.datastax.driver.core.{CloseFuture, Cluster, ResultSet, Session}
+import com.google.common.util.concurrent.{AsyncFunction, Futures, ListenableFuture, MoreExecutors}
+
+import scala.reflect.ClassTag
 
 package object api {
 
@@ -30,6 +33,33 @@ package object api {
 
   def apiInterpreter[F[_], A](a: A): (Kleisli[F, A, ?] ~> F) = new (Kleisli[F, A, ?] ~> F) {
     override def apply[B](fa: Kleisli[F, A, B]): F[B] = fa(a)
+  }
+
+  def kleisli[M[_], A, B](
+      f: A => M[B])(implicit ME: MonadError[M, Throwable], TAG: ClassTag[A]): Kleisli[M, A, B] =
+    Kleisli { (a: A) =>
+      Option(a)
+        .map(f)
+        .getOrElse(ME.raiseError(
+          new IllegalArgumentException(s"Instance of class ${TAG.runtimeClass.getName} is null")))
+    }
+
+  def closeFuture2unit[M[_], A](f: A => CloseFuture)(
+      implicit H: ListenableFuture[?] ~> M,
+      ME: MonadError[M, Throwable],
+      TAG: ClassTag[A]): Kleisli[M, A, Unit] = {
+
+    def listenableFuture(a: A): ListenableFuture[Unit] = Futures.transformAsync(
+      f(a),
+      new AsyncFunction[Void, Unit] {
+        override def apply(input: Void): ListenableFuture[Unit] =
+          Futures.immediateFuture((): Unit)
+      },
+      MoreExecutors.directExecutor()
+    )
+
+    kleisli(c => H(listenableFuture(c)))
+
   }
 
 }
