+++ core/src/test/scala/handlers/ClusterAPIHandlerSpec.scala
@@ -40,9 +40,10 @@ class ClusterAPIHandlerSpec
 
   import cats.instances.future._
   import freestyle.async.implicits._
-  import freestyle.cassandra.handlers.implicits._
+  import freestyle.asyncGuava.implicits._
+  import freestyle.cassandra.implicits._
   import TestUtils._
-  val handler: ClusterAPIHandler[Future] = clusterAPIHandler[Future]
+  val handler: ClusterAPIHandler[Future] = new ClusterAPIHandler[Future]
 
   "ListenableFutureHandler" should {
