+++ project/plugins.sbt
@@ -1,3 +1,3 @@
 resolvers += Resolver.sonatypeRepo("releases")
-addSbtPlugin("io.frees" % "sbt-freestyle" % "0.3.2")
+addSbtPlugin("io.frees" % "sbt-freestyle" % "0.3.9")
 addSbtPlugin("com.47deg" % "sbt-embedded-cassandra" % "0.0.4")
